generate json response in a bundle
